feat: update interface
