--- src/redux/reducers/editorReducer/dslReducer/app-state.tsx
@@ -1,7 +1,0 @@
-import { DslState } from "./dsl-state"
-import { MenuState } from "./menu-state"
-
-export interface AppState {
-  dslState: DslState
-  menuState: MenuState
-}
