--- src/redux/reducers/editorReducer/dslReducer/menu-state.tsx
@@ -1,5 +1,0 @@
-import { DslNode } from "./dsl-state"
-
-export interface MenuState {
-  current: DslNode | null
-}
